MOD : IStorageService moved to Application layer
IStorageService is an abstraction on which the Application layer depends, that is implemented in the Infrastructure layer.

Since it's the web layer that plug concrete infra implementations to application/domain abstraction, infra layer can depend on the application layer.

I have moved the IStorageService inside the application layer, and inverted the dependency between infra and application

Resolves #170